% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SphRankSign.R
\name{Fn}
\alias{Fn}
\title{Empirical distribution function for spherical data}
\usage{
Fn(data, nS, nR, n0)
}
\arguments{
\item{data}{spherical data}

\item{nS}{number of signs}

\item{nR}{number of ranks}

\item{n0}{number of sample points transported to the axis}
}
\value{
empirical distribution, axis
}
\description{
This function calculates the empirical distribution function for spherical data defined by optimal transport. The axis around which the uniform grids with rotational-symmetry is generated is also given in the results.
}
\keyword{OT}
\keyword{distribution}
\keyword{spherical}
