% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{opt_coupling}
\alias{opt_coupling}
\title{Compute the empirical distribution using Rcpp}
\usage{
opt_coupling(G, v)
}
\arguments{
\item{G}{the nxn matrix of transport cost(i.e. Riemannian distance)}

\item{v}{a integer vector with length n}
}
\value{
the optimal coupling expressed by permutation of 1..n
}
\description{
Find optimal coupling using KM algorithm
}
