% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SphRankSign.R
\name{RankSign}
\alias{RankSign}
\title{Ranks and signs for spherical data}
\usage{
RankSign(data, nS, nR, n0)
}
\arguments{
\item{data}{spherical data}

\item{nS}{number of signs}

\item{nR}{number of ranks}

\item{n0}{number of sample points transported to the axis}
}
\value{
ranks, signs, indices for data points that are transported to the axis, the axis
}
\description{
Calculates the ranks and signs based on empirical distribution for spherical data.
}
\keyword{rank}
\keyword{sign}
\keyword{spherical}
